engine: factorize value translation out of CommandPreprocessor
