+++ daml-lf/engine/src/main/scala/com/digitalasset/daml/lf/engine/Engine.scala
@@ -45,7 +45,7 @@ import com.digitalasset.daml.lf.speedy.{Command => SpeedyCommand}
   */
 final class Engine {
   private[this] val _compiledPackages = ConcurrentCompiledPackages()
-  private[this] val _commandTranslation = CommandPreprocessor(_compiledPackages)
+  private[this] val _commandTranslation = new CommandPreprocessor(_compiledPackages)
 
   /**
     * Executes commands `cmds` under the authority of `cmds.submitter` and returns one of the following:
